fix: refactor encrypt
