+++ src/utils/encrypt/index.ts
@@ -1,5 +1,6 @@
 import CryptoJS from "crypto-js";
 import { secretKeyPortalId } from "@config/environment";
+
 const secretKey = CryptoJS.enc.Hex.parse(secretKeyPortalId);
 const iv = CryptoJS.enc.Hex.parse("abcdef9876543210abcdef9876543210");
 
@@ -13,20 +14,8 @@ const encrypt = (data: string): string => {
     }
   } catch (error) {
     console.error("Encryption error:", error);
-
     return error instanceof Error ? error.message : "Unknown encryption error";
   }
 };
 
-const decrypt = (data: string) => {
-  try {
-    if (!data) return "";
-    const bytes = CryptoJS.AES.decrypt(data, secretKey, { iv: iv });
-    return bytes.toString(CryptoJS.enc.Utf8);
-  } catch (error) {
-    console.error("Decryption error:", error);
-    return "";
-  }
-};
-
-export { encrypt, decrypt };
+export { encrypt };
